subsys: usb: device_next: Fix unchecked return value
The return value of usbd_ep_enqueue() was previously unchecked in
cdc_ecm_send(), unlike other uses of the same API.

This could lead to a deadlock if the transfer fails and k_sem_take() is
still called.

This patch checks the return value, logs an error, skips the wait if the
enqueue fails, and ensures the buffer is properly unreferenced.

Fixes: CID 489224

Signed-off-by: Gaetan Perrot <gaetan.perrot@spacecubics.com>